Fix type casting in ternary and use signed timestamp for CRT hits in CAF

+++ sbncode/CAFMaker/FillReco.cxx
@@ -21,7 +21,7 @@ namespace caf
                   bool use_ts0,
                   caf::SRCRTHit &srhit,
                   bool allowEmpty) {
-    srhit.time = (use_ts0 ? hit.ts0_ns : hit.ts1_ns) / 1000.;
+    srhit.time = (use_ts0 ? (float)hit.ts0_ns : hit.ts1_ns) / 1000.;
 
     srhit.position.x = hit.x_pos;
     srhit.position.y = hit.y_pos;
